Extract blocking-stdio setup into a shared module; cover wait-for-host-standby
setup-logger.ts already put stdout/stderr in blocking mode under NODE_ENV
development so the service entrypoints flush logs in real time instead of
at process exit. The dev-only `wait-for-host-standby` probe runs as a
separate ts-node script in the same piped topology and was missing that
treatment — its retry-attempt output buffered until teardown, so a CI
failure couldn't tell us when its puppeteer probes actually finished.

Move the side-effect block into `lib/unbuffer-stdio.ts` and import it from
both setup-logger and wait-for-host-standby, so a hung startup probe is
diagnosable in real time too.
